# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for multi-branch cafe chain inventory",
        "title": "Cafe Inventory Management API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/cafes": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cafes"
                ],
                "summary": "Get all cafes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by nearest station",
                        "name": "station",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max results to return (default 20)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Cafe"
                            }
                        }
                    },
                    "400": {
                        "description": "invalid limit",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cafes"
                ],
                "summary": "Create a cafe",
                "parameters": [
                    {
                        "description": "Cafe payload",
                        "name": "cafe",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CafeCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/cafes/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cafes"
                ],
                "summary": "Get a cafe by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cafe ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Cafe"
                        }
                    },
                    "400": {
                        "description": "invalid cafe id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "data not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cafes"
                ],
                "summary": "Delete a cafe",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cafe ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "invalid id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "data not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "cafe has dependent records",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cafes"
                ],
                "summary": "Update a cafe",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cafe ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Cafe payload",
                        "name": "cafe",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CafeUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Cafe"
                        }
                    },
                    "400": {
                        "description": "invalid request body or invalid cafe id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "data not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Cafe": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "closing_time": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "nearest_station": {
                    "type": "string"
                },
                "opening_time": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.CafeCreate": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "closing_time": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "nearest_station": {
                    "type": "string"
                },
                "opening_time": {
                    "type": "string"
                }
            }
        },
        "model.CafeUpdate": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "closing_time": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "nearest_station": {
                    "type": "string"
                },
                "opening_time": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}